chore: Add classNames utility function
This commit adds the `classNames` utility function to the `src/lib/utils/classNames.ts` file. The `classNames` function is imported from the `clsx` package and uses the `twMerge` function from the `tailwind-merge` package to merge multiple class values into a single string. This utility function will be useful for generating dynamic class names in the project.