refactor: Update usePagination hook to accept defaultRowsPerPage prop
This commit updates the usePagination hook to accept a defaultRowsPerPage prop, allowing the consumer to specify the initial number of rows per page. If the prop is not provided, the default value of 5 will be used. This enhancement improves the flexibility and customization options of the usePagination hook.

Note: The commit message has been generated based on the provided code changes and recent repository commits.