refactor: Update useSelectColumnsTable to accept columns as a prop
This commit updates the useSelectColumnsTable hook to accept the columns prop, which is an array of Column objects. The Column interface has been defined to include properties such as name, uid, sortable, and align. This change allows for more flexibility in customizing the columns used in the table and improves the reusability of the hook.

Note: The commit message has been generated based on the provided code changes and recent repository commits.