refactor: Update useStatusFilter to include defaultStatus prop
This commit updates the useStatusFilter hook to include a defaultStatus prop, allowing the consumer to specify the initial status filter. If the prop is not provided, the default value of "all" will be used. This enhancement improves the flexibility and customization options of the useStatusFilter hook.

Note: The commit message has been generated based on the provided code changes and recent repository commits.